SCKAN-196 feat: Add menus to the graph diagram
